Conflicts:
	src/main/java/dal/admin/ImageStore.java
	src/test/java/test/StoreBlockingTest.java

+++ src/main/java/dal/admin/IImageStore.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 public interface IImageStore {
 	public boolean insert(Image img);
 	public ArrayList<Image> getLast(int numberOfRows);
-	public boolean block(int id);
-	public boolean unBlock(int id);
+	public boolean block(long id);
+	public boolean unBlock(long id);
 
 }
